meson: back-port support for fullcpp option
For cross-compiles nrfcxx builds this with fullcpp disabled and static
non-pic libraries.  However, for host builds for unit testing it needs a
pic library to link with the stub nrfcxx library.  That's controlled
through fullcpp, so we need the flag even though none of the
corresponding capability is in the public release.

As there is no implementation support for fullcpp capabilities the
preprocessor definition remains hard-coded to false.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>